Simplify runtime library to two headers (merge span_eigen.h into span.h)
