+++ components/runtime/CMakeLists.txt
@@ -1,7 +1,6 @@
 # Create a target for the runtime library (a header only target).
 set(RUNTIME_INTERFACE_NAME wf_runtime)
-set(RUNTIME_SOURCES wrenfold/span.h wrenfold/span_detail.h
-                    wrenfold/span_eigen.h)
+set(RUNTIME_SOURCES wrenfold/span.h wrenfold/span_detail.h)
 add_library(${RUNTIME_INTERFACE_NAME} INTERFACE ${RUNTIME_SOURCES})
 target_include_directories(
   ${RUNTIME_INTERFACE_NAME}
