+++ examples/bspline/bspline_test.cc
@@ -8,7 +8,7 @@
 #include "wf_test_support/test_macros.h"
 
 #define WF_SPAN_EIGEN_SUPPORT
-#include "wrenfold/span_eigen.h"
+#include "wrenfold/span.h"
 
 #include "bspline_numerical.h"
