+++ examples/quaternion_interpolation/quat_interpolation_test.cc
@@ -9,7 +9,7 @@
 #include "quat_interpolation_expressions.h"
 
 #define WF_SPAN_EIGEN_SUPPORT
-#include "wrenfold/span_eigen.h"
+#include "wrenfold/span.h"
 
 #include "generated.h"
